# application/server/openapi.yaml
openapi: 3.0.3
info:
  title: Web Speed Hackathon 2026 API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /crok/suggestions:
    get:
      tags:
        - Crok
      summary: チャット候補取得
      operationId: getCrokSuggestions
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: 検索クエリ。指定するとサーバー側で形態素解析+BM25によるフィルタリングを行い上位10件を返す。未指定時は全候補を返す。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - suggestions
                properties:
                  suggestions:
                    type: array
                    items:
                      type: string
                  queryTokens:
                    type: array
                    items:
                      type: string
                    description: 形態素解析で抽出された内容語トークン（qパラメータ指定時のみ）
  /crok/sentiment:
    get:
      tags:
        - Crok
      summary: テキスト感情分析
      operationId: analyzeSentiment
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
          description: 感情分析対象のテキスト
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - score
                  - label
                properties:
                  score:
                    type: number
                    description: 感情スコア（正=ポジティブ、負=ネガティブ）
                  label:
                    type: string
                    enum:
                      - positive
                      - negative
                      - neutral
  /crok:
    get:
      tags:
        - Crok
      summary: AI チャット（SSE ストリーミング）
      operationId: chatWithCrok
      description: Server-Sent Events で AI レスポンスを1文字ずつストリーミング返却する。
      security:
        - session: []
      parameters:
        - name: prompt
          in: query
          required: true
          schema:
            type: string
          description: ユーザーの質問
      responses:
        "200":
          description: OK
          content:
            text/event-stream:
              schema:
                description: |
                  各チャンクは SSE イベントとして送信される
                type: object
                properties:
                  text:
                    type: string
                  done:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: connect.sid
      description: express-session によるセッション Cookie
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...